修复我的账单页面收入、支出、结余计算精度的问题

+++ bookkeeping/bookkeeping/Classes/Modules/Bill（账单）/View/BillTable.h
@@ -12,8 +12,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface BillTable : UITableView
 
-@property (nonatomic, assign) NSInteger income;
-@property (nonatomic, assign) NSInteger pay;
+@property (nonatomic, assign) CGFloat income;
+@property (nonatomic, assign) CGFloat pay;
 @property (nonatomic, strong) NSMutableArray *models;
 
 @end
